requirements.txt
streamlit==1.55.0
pandas==2.3.3
numpy==2.4.2
plotly==6.5.2
pyarrow==23.0.1
python-dotenv==1.2.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.49
